Robin, We need to change the name of this Project to OCR-SOM
--
Changes:
Created the whole model for the SOM to Work.
The user will draw a letter in a small place in the view and click "ADD", the program will ask him to say what letter was that he drawn and it will be added to an arrayList.
And he can keep adding several of them.

When he finishes, he can click in train and the train method will start.
Its not optimized.Because im unsure it will work at the first time without bugs..

But all Classes are there.
--